Removed deprection warnings for Rails 5.1
The behavior of `attribute_changed?` inside of after callbacks will be
changing in the next version of Rails. The new return value will reflect
the behavior of calling the method after `save` returned (e.g. the
opposite of what it returns now). To maintain the current behavior, use
`saved_change_to_attribute?` instead.